@@ -1,84 +1,68 @@
 {
-  "name": "localize-router",
-  "version": "1.0.0-rc.4",
-  "description": "An implementation of routes localization for Angular",
+  "name": "localize-router6",
+  "version": "0.0.1",
   "scripts": {
-    "test": "karma start",
-    "test-watch": "karma start --singleRun=false --autoWatch=true",
-    "commit": "npm run prepublish && npm test",
-    "prepublish": "ngc && npm run build",
-    "prepare": "ngc && npm run build",
-    "build": "webpack"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Greentube/localize-router.git"
-  },
-  "keywords": [
-    "angular",
-    "angular2",
-    "translate",
-    "localize",
-    "routes",
-    "i18n"
-  ],
-  "author": "Miroslav Jonas",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Greentube/localize-router/issues"
-  },
-  "main": "bundles/localize-router.umd.js",
-  "module": "index.js",
-  "typings": "index.d.ts",
-  "homepage": "https://github.com/Greentube/localize-router",
+  "private": true,
   "peerDependencies": {
-    "@angular/core": ">=4.0.0",
-    "@angular/common": ">=4.0.0",
-    "@angular/router": ">=4.0.0",
-    "rxjs": "^5.0.0",
-    "@ngx-translate/core": ">=8.0.0"
+    "@ngx-translate/core": ">=10.0.2"
   },
   "devDependencies": {
-    "@angular/animations": "~4.3.3",
-    "@angular/common": "~4.3.3",
-    "@angular/compiler": "~4.3.3",
-    "@angular/compiler-cli": "~4.3.3",
-    "@angular/core": "~4.3.3",
-    "@angular/http": "~4.3.3",
-    "@angular/platform-browser": "~4.3.3",
-    "@angular/platform-browser-dynamic": "~4.3.3",
-    "@angular/platform-server": "~4.3.3",
-    "@angular/router": "~4.3.3",
-    "@ngx-translate/core": "8.0.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/animations": "~6.0.3",
+    "@angular/cli": "~6.0.8",
+    "@angular/common": "~6.0.3",
+    "@angular/compiler": "~6.0.3",
+    "@angular/compiler-cli": "~6.0.3",
+    "@angular/core": "~6.0.3",
+    "@angular/http": "~6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@angular/platform-browser": "~6.0.3",
+    "@angular/platform-browser-dynamic": "~6.0.3",
+    "@angular/platform-server": "~6.0.3",
+    "@angular/router": "~6.0.3",
+    "@ngx-translate/core": "10.0.2",
     "@types/es6-promise": "0.0.32",
     "@types/hammerjs": "2.0.34",
-    "@types/jasmine": "2.5.47",
-    "@types/node": "7.0.22",
+    "@types/jasmine": "^2.8.8",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "~8.9.4",
     "awesome-typescript-loader": "3.1.3",
-    "codelyzer": "4.0.1",
+    "codelyzer": "~4.2.1",
     "concurrently": "3.1.0",
     "core-js": "2.4.1",
     "istanbul-instrumenter-loader": "~2.0.0",
-    "jasmine-core": "2.6.2",
-    "karma": "1.7.0",
-    "karma-chrome-launcher": "2.1.1",
-    "karma-coverage": "1.1.1",
-    "karma-jasmine": "1.1.0",
-    "karma-mocha-reporter": "2.2.3",
-    "karma-remap-coverage": "0.1.5",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.3",
+    "jasmine-core": "^3.1.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^2.0.5",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage": "^1.1.2",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.2.0",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-remap-coverage": "^0.1.5",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^3.0.0",
     "loader-utils": "1.1.0",
+    "protractor": "~5.3.0",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.0",
+    "rxjs": "6.2.2",
     "source-map-loader": "0.2.1",
     "ts-helpers": "1.1.2",
-    "tslint": "5.8.0",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
     "tslint-eslint-rules": "^4.1.1",
     "tslint-loader": "3.5.3",
-    "typescript": "2.3.3",
-    "webpack": "2.6.0",
-    "webpack-node-externals": "^1.5.4",
-    "zone.js": "0.8.11"
+    "typescript": "~2.7.2",
+    "webpack": "^4.16.2",
+    "webpack-node-externals": "^1.7.2",
+    "zone.js": "0.8.26"
   }
 }
